+++ encoding/decoder_test.go
@@ -446,6 +446,20 @@ func TestDecodeMapIntKeys(t *testing.T) {
 	}
 }
 
+func TestDecodeCompoundKey(t *testing.T) {
+	input := map[string]interface{}{"id": []string{"1", "2"}, "err_a[]": "3", "err_b[": "4", "err_c]": "5"}
+	want := Compound{"1", "2", "3", "4", "5"}
+
+	out := Compound{}
+	err := Decode(&out, input)
+	if err != nil {
+		t.Errorf("got error %v, expected nil", err)
+	}
+	if !jsonEqual(out, want) {
+		t.Errorf("got %q, want %q", out, want)
+	}
+}
+
 func jsonEqual(a, b interface{}) bool {
 	// First check using reflect.DeepEqual
 	if reflect.DeepEqual(a, b) {
@@ -463,4 +477,33 @@ func jsonEqual(a, b interface{}) bool {
 	}
 
 	return bytes.Compare(ba, bb) == 0
+}
+
+func TestMergeStruct(t *testing.T) {
+	var dst struct {
+		Field        string
+		AnotherField string
+	}
+	dst.Field = "change me"
+	dst.AnotherField = "don't blank me"
+	err := Merge(&dst, map[string]interface{}{"Field": "Changed!"})
+	if err != nil {
+		t.Error("Cannot merge:", err)
+	}
+	if dst.AnotherField == "" {
+		t.Error("Field has been wiped")
+	}
+}
+
+func TestMergeMap(t *testing.T) {
+	var dst = make(map[string]string)
+	dst["field"] = "change me"
+	dst["another_field"] = "don't blank me"
+	err := Merge(&dst, map[string]interface{}{"field": "Changed!"})
+	if err != nil {
+		t.Error("Cannot merge:", err)
+	}
+	if dst["another_field"] == "" {
+		t.Error("Field has been wiped")
+	}
 }