+++ encoding/encoder_test.go
@@ -347,3 +347,63 @@ func TestReferenceFieldArray(t *testing.T) {
 		t.Errorf("got %q, want %q", out, want)
 	}
 }
+
+func TestEncodeBytes(t *testing.T) {
+	type BytesStruct struct {
+		A []byte
+		B [1]byte
+	}
+
+	input := BytesStruct{[]byte("A"), [1]byte{'B'}}
+	want := map[string]interface{}{
+		"A": map[string]interface{}{"$reql_type$": "BINARY", "data": "QQ=="},
+		"B": map[string]interface{}{"$reql_type$": "BINARY", "data": "Qg=="},
+	}
+
+	out, err := Encode(input)
+	if err != nil {
+		t.Errorf("got error %v, expected nil", err)
+	}
+	if !jsonEqual(out, want) {
+		t.Errorf("got %q, want %q", out, want)
+	}
+}
+
+type Compound struct {
+	PartA string `gorethink:"id[0]"`
+	PartB string `gorethink:"id[1]"`
+	ErrA  string `gorethink:"err_a[]"`
+	ErrB  string `gorethink:"err_b["`
+	ErrC  string `gorethink:"err_c]"`
+}
+
+func TestEncodeCompound(t *testing.T) {
+	input := Compound{"1", "2", "3", "4", "5"}
+	want := map[string]interface{}{"id": []string{"1", "2"}, "err_a[]": "3", "err_b[": "4", "err_c]": "5"}
+
+	out, err := Encode(input)
+	if err != nil {
+		t.Errorf("got error %v, expected nil", err)
+	}
+	if !jsonEqual(out, want) {
+		t.Errorf("got %q, want %q", out, want)
+	}
+}
+
+type CompoundRef struct {
+	PartA string `gorethink:"id[0]"`
+	PartB *RefB  `gorethink:"id[1],reference" gorethink_ref:"id"`
+}
+
+func TestEncodeCompoundRef(t *testing.T) {
+	input := CompoundRef{"1", &RefB{"2", "Name"}}
+	want := map[string]interface{}{"id": []string{"1", "2"}}
+
+	out, err := Encode(input)
+	if err != nil {
+		t.Errorf("got error %v, expected nil", err)
+	}
+	if !jsonEqual(out, want) {
+		t.Errorf("got %q, want %q", out, want)
+	}
+}