+++ session_test.go
@@ -130,7 +130,7 @@ func (s *RethinkSuite) TestSessionConnectDatabase(c *test.C) {
 
 	_, err = Table("test2").Run(session)
 	c.Assert(err, test.NotNil)
-	c.Assert(err.Error(), test.Equals, "gorethink: Database `test2` does not exist. in: \nr.Table(\"test2\")")
+	c.Assert(err.Error(), test.Equals, "gorethink: Database `test2` does not exist. in:\nr.Table(\"test2\")")
 
 	session.Use("test3")
 	c.Assert(session.Database(), test.Equals, "test3")